+++ src/main/java/org/folio/search/service/converter/preprocessor/extractor/impl/ClassificationResourceExtractor.java
@@ -12,6 +12,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import lombok.extern.log4j.Log4j2;
+import org.apache.commons.lang3.StringUtils;
 import org.folio.search.domain.dto.ResourceEvent;
 import org.folio.search.domain.dto.TenantConfiguredFeature;
 import org.folio.search.model.types.ResourceType;
@@ -57,7 +58,7 @@ public class ClassificationResourceExtractor extends ChildResourceExtractor {
       return Collections.emptyMap();
     }
     var classificationNumber = prepareForExpectedFormat(entityProperties.get(CLASSIFICATION_NUMBER_FIELD), 50);
-    if (classificationNumber.isEmpty()) {
+    if (StringUtils.isBlank(classificationNumber)) {
       return Collections.emptyMap();
     }