+++ src/test/java/org/folio/search/service/converter/preprocessor/extractor/impl/SubjectResourceExtractorTest.java
@@ -41,9 +41,24 @@ class SubjectResourceExtractorTest extends ChildResourceExtractorTestBase {
     persistChildrenTest(extractor, repository, subjectsBodySupplier());
   }
 
+  @Test
+  void shouldNotPersistEmptySubject() {
+    when(configService.isEnabled(TenantConfiguredFeature.BROWSE_SUBJECTS)).thenReturn(true);
+    shouldNotPersistEmptyChildrenTest(extractor, repository, emptySubjectsBodySupplier());
+  }
+
   private static Supplier<Map<String, Object>> subjectsBodySupplier() {
     return () -> Map.of(SUBJECTS_FIELD, List.of(Map.of(
       SUBJECT_VALUE_FIELD, RandomStringUtils.insecure().nextAlphanumeric(260),
+      AUTHORITY_ID_FIELD, UUID.randomUUID().toString(),
+      SUBJECT_SOURCE_ID_FIELD, UUID.randomUUID().toString(),
+      SUBJECT_TYPE_ID_FIELD, UUID.randomUUID().toString()
+    )));
+  }
+
+  private static Supplier<Map<String, Object>> emptySubjectsBodySupplier() {
+    return () -> Map.of(SUBJECTS_FIELD, List.of(Map.of(
+      SUBJECT_VALUE_FIELD, "        ",
       AUTHORITY_ID_FIELD, UUID.randomUUID().toString(),
       SUBJECT_SOURCE_ID_FIELD, UUID.randomUUID().toString(),
       SUBJECT_TYPE_ID_FIELD, UUID.randomUUID().toString()